requirements.txt
crypto-data-downloader
Pillow
